requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0

# Audio Processing
soundfile==0.12.1
librosa==0.10.1
pydub==0.25.1
numpy==1.24.3
scipy==1.11.4

# MIDI Processing
mido==1.3.2
python-rtmidi==1.5.6

# AI/ML Integration
openai==1.6.1
anthropic==0.8.1
torch==2.1.2
transformers==4.36.2

# Music Theory & Analysis
music21==9.1.0
pretty_midi==0.2.10

# Database
SQLAlchemy==2.0.25
alembic==1.13.1

# Audio Effects & Synthesis
pedalboard==0.9.8

# Real-time Audio
pyaudio==0.2.14
python-sounddevice==0.4.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
click==8.1.7

# Data Validation
pydantic==2.5.3
marshmallow==3.20.1

# Task Queue (for background processing)
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
